# Conflicts:
#	.vs/Massacre_Project/FileContentIndex/29dc9490-05e6-470e-bca2-bbc414bad532.vsidx
#	.vs/Massacre_Project/FileContentIndex/b71f2adc-4417-422f-b64a-e2732171117e.vsidx
#	.vs/Massacre_Project/FileContentIndex/eeff42bc-9f10-4a47-b6d9-6d6353f7895c.vsidx
#	.vs/Massacre_Project/v17/fileList.bin
#	UserSettings/EditorUserSettings.asset
#	UserSettings/Layouts/default-2022.dwlt

+++ Assets/01.BSJ/02.Scripts/Monster/Combat/MonsterHealth.cs
@@ -3,17 +3,17 @@ using UnityEngine;
 [System.Serializable]
 public class MonsterHealth
 {
-    [SerializeField] private int _currentHealth;
-    [SerializeField] private int _maxHealth;
+    [SerializeField] private float _currentHealth;
+    [SerializeField] private float _maxHealth;
 
-    public MonsterHealth(int currentHealth, int maxHealth)
+    public MonsterHealth(float currentHealth, float maxHealth)
     {
         _currentHealth = currentHealth;
         _maxHealth = maxHealth;
     }
 
-    public int CurrentHealth { get => _currentHealth; set => _currentHealth = value; }
-    public int MaxHealth { get => _maxHealth; set => _maxHealth = value; }
+    public float  CurrentHealth { get => _currentHealth; set => _currentHealth = value; }
+    public float  MaxHealth { get => _maxHealth; set => _maxHealth = value; }
 
     public void InitializeHealth()
     {
